Removes dirty word parser from HexUtils and puts into the New Message Processor.
